+++ packages/runtime/adapters/eventstore-adapters/eventstore-postgresql/src/save-event.ts
@@ -2,10 +2,14 @@ import {
   ConcurrentError,
   InputEvent,
   EventstoreFrozenError,
+  SavedEvent,
+  threadArrayToCursor,
+  initThreadArray,
 } from '@resolve-js/eventstore-base'
 
 import { RESERVED_EVENT_SIZE, LONG_NUMBER_SQL_TYPE } from './constants'
 import { AdapterPool } from './types'
+import assert from 'assert'
 
 const saveEvent = async (
   {
@@ -16,7 +20,7 @@ const saveEvent = async (
     escape,
   }: AdapterPool,
   event: InputEvent
-): Promise<void> => {
+): Promise<string> => {
   while (true) {
     try {
       const serializedEvent = [
@@ -36,7 +40,7 @@ const saveEvent = async (
       const threadsTableAsId = escapeId(`${eventsTableName}-threads`)
       const eventsTableAsId = escapeId(eventsTableName)
 
-      await executeStatement(
+      const rows = (await executeStatement(
         `WITH "freeze_check" AS (
           SELECT 0 AS "freeze_zero" WHERE (
             (SELECT 1 AS "EventStoreIsFrozen")
@@ -71,7 +75,7 @@ const saveEvent = async (
             SELECT "threadId" FROM "vector_id" LIMIT 1
           )
           RETURNING *
-        ) INSERT INTO ${databaseNameAsId}.${eventsTableAsId}(
+        ), "insert_event" AS (INSERT INTO ${databaseNameAsId}.${eventsTableAsId}(
           "threadId",
           "threadCounter",
           "timestamp",
@@ -90,10 +94,22 @@ const saveEvent = async (
           ),
           ${serializedEvent},
           ${byteLength}
-        )`
-      )
+        )) SELECT "threadId", "threadCounter",
+        (CASE WHEN (SELECT "threadId" FROM "vector_id" LIMIT 1) = "threadId" THEN "threadCounter"+1
+          ELSE "threadCounter" END) AS "newThreadCounter"
+        FROM ${databaseNameAsId}.${threadsTableAsId}
+        ORDER BY "threadId" ASC`
+      )) as Array<{
+        threadId: SavedEvent['threadId']
+        newThreadCounter: SavedEvent['threadCounter']
+      }>
 
-      break
+      assert.strictEqual(rows.length, 256, 'Thread table must have 256 rows')
+      const threadCounters = initThreadArray()
+      for (const row of rows) {
+        threadCounters[row.threadId] = row.newThreadCounter
+      }
+      return threadArrayToCursor(threadCounters)
     } catch (error) {
       const errorMessage =
         error != null && error.message != null ? error.message : ''