+++ packages/tools/testing-tools/test/read-model.test.ts
@@ -1,6 +1,7 @@
 import givenEvents from '../src/index'
 import { Event, EventHandlerEncryptionContext } from '@resolve-js/core'
 import { TestReadModel } from '../types/types'
+import { ambiguousEventsTimeErrorMessage } from '../src/constants'
 
 const ProjectionError = (function (this: Error, message: string): void {
   Error.call(this)
@@ -147,6 +148,107 @@ describe('sequence tests', () => {
       .query('all', {})
 
     expect(result?.items).toEqual(['test-1', 'test-2', 'test-3'])
+  })
+})
+
+describe('givenEvents tests', () => {
+  const testId = 'root'
+
+  const readModel: TestReadModel = {
+    name: 'readModelSnapshotTimestamps',
+    projection: {
+      Init: async (store: any): Promise<any> => {
+        await store.defineTable('snapshot', {
+          indexes: { id: 'string' },
+          fields: ['items', 'timestamp'],
+        })
+        await store.insert('snapshot', { id: testId, items: [] })
+      },
+      TEST: async (
+        store: any,
+        { timestamp, payload: { item } }
+      ): Promise<any> => {
+        const { items } = await store.findOne('snapshot', {
+          id: testId,
+        })
+        await store.update(
+          'snapshot',
+          { id: testId },
+          {
+            $set: {
+              items: [
+                ...items,
+                {
+                  value: item,
+                  timestamp,
+                },
+              ],
+            },
+          }
+        )
+      },
+    },
+    resolvers: {
+      all: async (store: any): Promise<any> => {
+        return await store.findOne('snapshot', { id: testId })
+      },
+    },
+  }
+
+  test('should match snapshot', async () => {
+    const result: any = await givenEvents([
+      { aggregateId: 'id1', type: 'TEST', payload: { item: 'test-1' } },
+      { aggregateId: 'id2', type: 'TEST', payload: { item: 'test-2' } },
+      { aggregateId: 'id3', type: 'TEST', payload: { item: 'test-3' } },
+    ])
+      .readModel(readModel)
+      .query('all', {})
+
+    expect(result?.items).toMatchSnapshot()
+  })
+
+  test('should match snapshot with specified timestamp', async () => {
+    const result: any = await givenEvents([
+      {
+        aggregateId: 'id1',
+        type: 'TEST',
+        payload: { item: 'test-1' },
+        timestamp: 10,
+      },
+      {
+        aggregateId: 'id2',
+        type: 'TEST',
+        payload: { item: 'test-2' },
+        timestamp: 20,
+      },
+      {
+        aggregateId: 'id3',
+        type: 'TEST',
+        payload: { item: 'test-3' },
+        timestamp: 30,
+      },
+    ])
+      .readModel(readModel)
+      .query('all', {})
+
+    expect(result?.items).toMatchSnapshot()
+  })
+
+  test('should throw error', async () => {
+    await expect(
+      givenEvents([
+        { aggregateId: 'id1', type: 'TEST', payload: { item: 'test-1' } },
+        {
+          aggregateId: 'id2',
+          type: 'TEST',
+          payload: { item: 'test-2' },
+          timestamp: 20,
+        },
+        { aggregateId: 'id3', type: 'TEST', payload: { item: 'test-3' } },
+      ])
+        .readModel(readModel)
+        .query('all', {})
+    ).rejects.toThrow(ambiguousEventsTimeErrorMessage)
   })
 })